Refactored the eval suite
