+++ tests/dpsdgym/artifact.json
@@ -1,0 +1,1 @@
+{"mushroom": {"data": {}, "target": "edible", "name": "mushroom", "mwem": {"0.01": {}, "0.1": {}, "1.0": {}, "9.0": {}, "45.0": {}, "95.0": {}}, "quail_synth": {"0.01": {}, "0.1": {}, "1.0": {}, "9.0": {}, "45.0": {}, "95.0": {}}}}
